# specs/001-/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Pomodoro Task Management API
  description: REST API for cross-platform Pomodoro task and time management application
  version: 1.0.0
servers:
  - url: https://api.pomodoro-app.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register new user account
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      summary: Authenticate user
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  # Tasks
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        access_token:
          type: string
        expires_in:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        preferences:
          type: object
        created_at:
          type: string
          format: date-time
...